Hotfix release: freezing browserSync version
BrowserSync has been frozen to 2.2.0. If a newer version is
installed, the HTML snippet has to be modified, causing a few
problems.

Found this issue when debugging an error. An HTML snippet must
be placed on the bottom of the site. That snippet was not copied
properly and a few characters were lost, causing a JavaScript
error. This one is fixed too.

Sorry, I'm not used to copy+paste. Honestly, it will happen
again.